requirements.txt
icalevents
pyyaml
dhooks